+++ packages/timebar/src/timebar.js
@@ -220,6 +220,7 @@ class Timebar extends Component {
       absoluteStart,
       bookmarkStart,
       bookmarkEnd,
+      bookmarkPlacement,
       enablePlayback,
       minimumRange,
       minimumRangeUnit,
@@ -332,6 +333,7 @@ class Timebar extends Component {
             onBookmarkChange={this.props.onBookmarkChange}
             bookmarkStart={bookmarkStart}
             bookmarkEnd={bookmarkEnd}
+            bookmarkPlacement={bookmarkPlacement}
             showLastUpdate={this.props.showLastUpdate}
           />
         </div>
@@ -347,6 +349,7 @@ Timebar.propTypes = {
   children: PropTypes.func,
   bookmarkStart: PropTypes.string,
   bookmarkEnd: PropTypes.string,
+  bookmarkPlacement: PropTypes.string,
   onMouseLeave: PropTypes.func,
   onMouseMove: PropTypes.func,
   onBookmarkChange: PropTypes.func,
@@ -369,6 +372,7 @@ Timebar.defaultProps = {
   children: () => {},
   onMouseLeave: () => {},
   onMouseMove: () => {},
+  bookmarkPlacement: 'top',
   onBookmarkChange: () => {},
   minimumRange: null,
   minimumRangeUnit: 'day',