+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,10 +1,17 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.32
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/timebar@1.3.1
+
 ## 0.7.31
 
 ### Patch Changes
 
-- fix crash on api auth errors
+- @globalfishingwatch/timebar@1.3.0
 
 ## 0.7.30