+++ packages/timebar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/timebar
 
+## 1.3.1
+
+### Patch Changes
+
+- remove missing log
+
+## 1.3.0
+
+### Minor Changes
+
+- "move timebar bookmark to bottom"
+
 ## 1.2.1
 
 ### Patch Changes